fix overflow for softmax kernel

+++ fastfold/model/fastnn/kernel/cuda_native/csrc/softmax_cuda.cpp
@@ -1,18 +1,18 @@
 #include <torch/extension.h>
 
-at::Tensor softmax(at::Tensor input, int rows, int cols);
-at::Tensor softmax_gradient(at::Tensor d_output, at::Tensor output, int rows, int cols);
+at::Tensor softmax(at::Tensor input, long long rows, long long cols);
+at::Tensor softmax_gradient(at::Tensor d_output, at::Tensor output, long long rows, long long cols);
 
-at::Tensor fused_scale_mask_softmax_forward(at::Tensor input, at::Tensor mask, int rows, int cols,
+at::Tensor fused_scale_mask_softmax_forward(at::Tensor input, at::Tensor mask, long long rows, long long cols,
                                             float scale);
 at::Tensor fused_scale_mask_softmax_backward(at::Tensor d_output, at::Tensor input, at::Tensor mask,
-                                             int rows, int cols, float scale);
+                                             long long rows, long long cols, float scale);
 
 at::Tensor fused_scale_mask_bias_softmax_forward(at::Tensor input, at::Tensor mask, at::Tensor bias,
-                                                 int rows, int cols, float scale);
+                                                 long long rows, long long cols, float scale);
 at::Tensor fused_scale_mask_bias_softmax_backward(at::Tensor d_output, at::Tensor input,
-                                                  at::Tensor mask, at::Tensor bias, int rows,
-                                                  int cols, float scale);
+                                                  at::Tensor mask, at::Tensor bias, long long rows,
+                                                  long long cols, float scale);
 
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
     m.def("forward", &softmax, "Softmax forward (CUDA)");
